% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{association}
\alias{association}
\title{association}
\usage{
association(x, target = NULL, threshold = 0, bigMatrix = 10)
}
\arguments{
\item{x}{dataframe containing data to be analyzed. Categorical variables
must be factors.}

\item{target}{optional character string containing the column name
for the target variable.  If this is provided, the target variable
will be excluded from the analysis. Submit the target variable
when analyzing associations only among predictors.}

\item{threshold}{numeric indicating the minimum association threshold
to be reported in the association table.}

\item{bigMatrix}{Numeric indicating the number of terms, above which,
the terms are replaced by short labels for association plotting.}
}
\value{
List containing the association matrix and pairwise association
table. If there are greater than 10 variables, the variable names
are replaced by a 2-3 character label for plotting purposes.
}
\description{
\code{association} Computes and summarizes an association matrix
}
\details{
Creates a association matrix and summarizes pairwise associations
in a association table.
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{association Functions}
