% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assocMatrix.R
\name{assocMatrix}
\alias{assocMatrix}
\title{assocMatrix}
\usage{
assocMatrix(x, vars, target)
}
\arguments{
\item{x}{Dataframe to be summarized}

\item{vars}{Character vector containing the names of the variables to analyse}
}
\value{
A matrix of associations between pairs of nominal variables
}
\description{
\code{assocMatrix} Creates a matrix of associations between nominal variables.
}
\seealso{
Other Analysis Functions: \code{\link{multiSLM}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Analysis Functions}
