% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{correlation}
\usage{
correlation(x, target = NULL, threshold = 0, bigMatrix = 10)
}
\arguments{
\item{x}{dataframe containing data to be analyzed}

\item{target}{optional character string containing the column name
for the target variable.  If this is provided, the target variable
will be excluded from the analysis. Submit the target variable
when analyzing correlations only among predictors.}

\item{threshold}{numeric indicating the minimum correlation threshold
to be reported in the correlation table.}

\item{bigMatrix}{Numeric indicating the number of terms above which
the terms are replaced with short labels for plotting.}
}
\value{
List containing the correlation matrix and pairwise correlation
table. If there are greater than 10 variables, the variable names
are replaced by a 2-3 character label for plotting purposes.
}
\description{
\code{correlation} Computes and summarizes a correlation matrix
}
\details{
Creates a correlation matrix and summarizes pairwise correlations
in a correlation table.
}
\seealso{
Other Correlation Functions: \code{\link{corrTest}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Correlation Functions}
