% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrTest.R
\name{corrTest}
\alias{corrTest}
\title{corrTest}
\usage{
corrTest(x, target, method = "pearson")
}
\arguments{
\item{x}{Dataframe to be analyzed}

\item{target}{Character string containig the column name of target variable.}

\item{method}{The correlation formula to use. Valid values include c("pearson", "kendall", "spearman")}
}
\value{
Dataframe of correlations between predictors and the target variable
}
\description{
\code{corrTest} Computes correlations between numeric predictors and a target variable
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Correlation Functions}
\concept{EDA Functions}
