% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiSLM.R
\name{multiSLM}
\alias{multiSLM}
\title{multiSLM}
\usage{
multiSLM(x, vars, target)
}
\arguments{
\item{x}{Dataframe to be summarized}

\item{vars}{Character vector containing the names of the variables to analyse}

\item{target}{Column name of target variable.}
}
\value{
A list containing a list of model information by predictor and a
dataframe summarizing regression statistics by predictor. The model information
is comprised of a list containing the model object, the regression results,
and the regression statistics.
}
\description{
\code{multiSLM} Performs simple linear regression on multiple variables.
}
\details{
Performs simple linear regression on multiple variables
vis-a-vis the target variable and returns a list containing the models,
the regression results and the statistics. The regression statisics for
each variable are summarized in descending order of r squared.
}
\seealso{
Other Analysis Functions: \code{\link{assocMatrix}}
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
\concept{Analysis Functions}
