% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{outliers}
\alias{outliers}
\title{outliers}
\usage{
outliers(x, target, observations)
}
\arguments{
\item{x}{An LM model for evaluation.}

\item{target}{Character string containing the column name of outcome or target variable.}

\item{observations}{Numeric vector of outlier observations}
}
\value{
List containing two objects. The first is a dataframe containing
a summary of the adjusted R2 performance with and without each of the
designated observations.  The second object is list containing
the observation number, the LM object with the observation removed
and the regression statistics for the lm objects.
}
\description{
\code{outliers} Statistics and plots to evaluate influence of outliers.
}
\details{
Regression statistics are computed and plots are rendered with and without
each outlier. Regression results with all designated outliers removed
are also provided.
}
\author{
John James, \email{jjames@datasciencesalon.org}
}
