% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printer.R
\name{printer}
\alias{printer}
\title{Silly Printer Function}
\usage{
printer(r, x)
}
\arguments{
\item{r}{A vector of length $n$}

\item{y}{A vector of length $n$}
}
\value{
The first 5 rows of the data frame as a tibble
}
\description{
Silly Printer Function
}
