% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdac_files_partial.R
\docType{data}
\name{sdac_files_partial}
\alias{sdac_files_partial}
\title{Switchboard Dialog Act Corpus Files - Partially curated}
\format{A data frame with 223,606 rows and 2 variables:
\describe{
  \item{doc_id}{ID for each conversation document}
  \item{text}{The actual dialog utterance}
}}
\source{
\url{https://catalog.ldc.upenn.edu/docs/LDC97S62/}
}
\usage{
sdac_files_partial
}
\description{
A dataset of the corpus files containing the 1,1150 conversations of 440 speakers of American
English.
}
\keyword{datasets}
