% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sms.R
\docType{data}
\name{sms}
\alias{sms}
\title{SMS Spam Collection v. 1}
\format{A data frame with 4,837 rows and 2 variables:
\describe{
  \item{sms_type}{Type of message; either spam or ham}
  \item{message}{The sms message}
}}
\source{
\url{http://www.dt.fee.unicamp.br/~tiago/smsspamcollection/smsspamcollection.zip}
}
\usage{
sms
}
\description{
The SMS Spam Collection v.1 is a public set of SMS labeled messages that
have been collected for mobile phone spam research. It has one collection
composed by 5,574 English, real and non-enconded messages, tagged according
being legitimate (ham) or spam.
}
\keyword{datasets}
